fix(refactored-structure): structure clean up remove unncessary things
